Finished Implementing the Closest Pair Algorithm Added a method to Edge to check if it has a point Added a method to Graph to copy another graph into this graph... for copying purposes Changed point to use double rather than int Setup the RobotTour controller class to run and test both algorithms

+++ src/chapter1/exercise26/Point.java
@@ -8,20 +8,20 @@ package chapter1.exercise26;
 public final class Point
 {
 
-    private final int x, y;
+    private final double x, y;
 
-    public Point(int x, int y)
+    public Point(double x, double y)
     {
         this.x = x;
         this.y = y;
     }
 
-    public int getX()
+    public double getX()
     {
         return this.x;
     }
 
-    public int getY()
+    public double getY()
     {
         return this.y;
     }
